sync 6208
git-svn-id: http://www.l2jtw.com:3690/svn/datapack@379 9bd44e3e-6552-0410-907e-a6ca40b856b6

+++ L2JTW_Datapack_CT2/data/scripts/handlers/admincommandhandlers/AdminCreateItem.java
@@ -56,7 +56,7 @@ public class AdminCreateItem implements IAdminCommandHandler
 					String id = st.nextToken();
 					int idval = Integer.parseInt(id);
 					String num = st.nextToken();
-					int numval = Integer.parseInt(num);
+					long numval = Long.parseLong(num);
 					createItem(activeChar, idval, numval);
 				}
 				else if (st.countTokens() == 1)
@@ -84,7 +84,7 @@ public class AdminCreateItem implements IAdminCommandHandler
 		return ADMIN_COMMANDS;
 	}
 	
-	private void createItem(L2PcInstance activeChar, int id, int num)
+	private void createItem(L2PcInstance activeChar, int id, long num)
 	{
 		L2Item template = ItemTable.getInstance().getTemplate(id);
 		if (template == null)
